Part 5 - Files structure Changes after review.
